requirements.txt
qrcode[pil]==8.2
python-telegram-bot==21.5
python-dotenv